+++ lib/apps/App.component.js
@@ -15,7 +15,7 @@ export default class App extends React.Component {
     return(
       <div className='dashboard-app'>
         <div className='dashboard-app-star'>
-          <figure className='star' onClick={ follow }/>
+          <figure className='star' onClick={ follow } />
         </div>
         <div className='dashboard-app-detail'>
           <header className='app-detail-header'>
