+++ lib/apps/apps.actions.js
@@ -27,8 +27,3 @@ export const Remove = "apps:remove"
 
 export const insert = (id, app) => ({ type: Insert, id, app })
 export const remove = (id) => ({ type: Remove, id })
-
-export const AddComponent = "apps:components:add"
-export const addComponent = (id, componentId) => (
-  { type: AddComponent, id, componentId }
-)
