+++ lib/components/components.sagas.js
@@ -18,7 +18,6 @@ function* fetchComponents({ appName }) {
 
   for (let component of components) {
     yield put(ComponentActions.insert(component.get('id'), component))
-    yield put(AppActions.addComponent(appName, component.get('id')))
   }
 }
 
@@ -32,7 +31,6 @@ function* getComponent({ id, appName }) {
   let component = ComponentEntity.Component.from(body, appName).toMap()
 
   yield put(ComponentActions.insert(component.get('id'), component))
-  yield put(AppActions.addComponent(appName, component.get('id')))
 }
 
 function* createComponent({ params }) {
@@ -46,7 +44,6 @@ function* createComponent({ params }) {
 
   if (response.status === 201) {
     yield put(ComponentActions.insert(component.get('id'), component))
-    yield put(AppActions.addComponent(appName, component.get('id')))
 
     yield fork(
       infoMessages,
