+++ lib/reducers.js
@@ -2,135 +2,19 @@ import { fromJS } from 'immutable'
 import { reducer as formReducer } from 'redux-form'
 import { combine } from './shared/combine.reducer'
 
-import * as AppBehavior from './apps/apps.behavior'
-import * as AppActions from './apps/apps.actions'
-
-import * as ComponentActions from './components/components.actions'
-import * as LayoutActions from './layouts/layouts.actions'
 import * as NotificationsActions from './notifications/notifications.actions'
 
+import apps from './apps/apps.reducers'
+import components from './components/components.reducers'
 import containers from './containers/containers.reducers'
 import instances from './instances/instances.reducers'
 import entrypoints from './entrypoints/entrypoints.reducers'
+import layouts from './layouts/layouts.reducers'
+import notifications from './notifications/notifications.reducers'
 import registries from './registries/registries.reducers'
 import releases from './releases/releases.reducers'
 import volumes from './volumes/volumes.reducers'
 import nodes from './nodes/nodes.reducers'
-
-function apps(state = fromJS({}), action) {
-  let app = state.get(action.id) || state.get(action.appName)
-  let componentIds = fromJS([])
-
-  if (app) { componentIds = app.get('components') || fromJS([]) }
-
-  switch (action.type) {
-    case AppActions.AddComponent:
-      return AppBehavior.addComponent(state, action.id, action.componentId)
-
-    case AppActions.Insert:
-      return AppBehavior.addApp(state, action.app)
-
-    case AppActions.Remove:
-      return state.delete(action.id)
-
-    case AppActions.SortComponents:
-      return AppBehavior.sortComponents(
-        state,
-        app.get('name'),
-        action.index,
-        action.componentId
-      )
-
-    case ComponentActions.Destroy:
-      return AppBehavior.removeComponent(state, app.get('name'), action.id)
-
-    default:
-      return state
-  }
-}
-
-function clouds(state = fromJS({}), action) {
-  let cloud = state.toList().first()
-  switch (action.type) {
-    case AppActions.Insert:
-      return state.set(
-        cloud.get('name'),
-        cloud.update(
-          'apps',
-          list => {
-            let name = action.app.get('name')
-
-            if (!list) { return fromJS([name]) }
-
-            return list.indexOf(name) >= 0 ? list : list.push(name)
-          }
-        )
-      )
-
-    case AppActions.Remove:
-      return state.set(
-        cloud.get('name'),
-        cloud.update(
-          'apps',
-          list => {
-            let index = list.indexOf(action.id)
-            return index >= 0 ? list.remove(index) : list
-          }
-        )
-      )
-
-    default:
-      return state
-  }
-}
-
-function components(state = fromJS({}), action) {
-  switch (action.type) {
-    case ComponentActions.Insert:
-      return state.set(action.component.get('id'), action.component)
-
-    case ComponentActions.Destroy:
-      return state.delete(action.id)
-
-    case ComponentActions.Update:
-      return state.set(
-        action.id,
-        state.get(action.id).merge(fromJS(action.params))
-      )
-
-    default:
-      return state
-  }
-}
-
-function layouts(state = fromJS({}), action) {
-  switch (action.type) {
-    case LayoutActions.Fade:
-      return state.set('faded', true)
-
-    case LayoutActions.Unfade:
-      return state.set('faded', false)
-
-    default:
-      return state
-  }
-}
-
-function meta(state = fromJS({}), action) { return state }
-
-function notifications(state = fromJS({}), action) {
-  const { id, level, message } = action
-  switch (action.type) {
-    case NotificationsActions.Add:
-      return state.set(id, fromJS({ level, message, id }))
-
-    case NotificationsActions.Remove:
-      return state.delete(id)
-
-    default:
-      return state
-  }
-}
 
 // Redux Form doesn't natively understand immutable things, so we do some
 // contortions around that here.  Basically what you're looking at is a snippet
@@ -176,7 +60,6 @@ function routing(state = fromJS({}), { type, payload }) {
 
 const manifest = {
   apps,
-  clouds,
   components,
   containers,
   entrypoints,
@@ -184,7 +67,6 @@ const manifest = {
   instances,
   layouts,
   notifications,
-  meta,
   nodes,
   releases,
   registries,
