+++ lib/selectors.js
@@ -80,10 +80,9 @@ export function createAppComponentsSelector(app) {
   return createSelector(
     getComponents,
     components => {
-      let names = app.get('components') || fromJS([])
-      let referenced = names.map(name => components.get(name))
-
-      return referenced.filter(component => !!component)
+      return components.filter(
+        component => component.get('id').includes(app.get('name'))
+      ).toList()
     }
   )
 }
